Clean up theme definitions in vanilla lib
